Clean up code, refactor models, update error-handling
