+++ lib/base/helpers/model_generator.dart
@@ -1,5 +1,8 @@
 import 'dart:math';
-import 'package:gocast_mobile/model/user_model.dart';
+import 'package:gocast_mobile/model/course/bookmark_model.dart';
+import 'package:gocast_mobile/model/course/course_model.dart';
+import 'package:gocast_mobile/model/user/user_model.dart';
+import 'package:gocast_mobile/model/user/user_settings_model.dart';
 
 class ModelGenerator {
   static User generateRandomUser() {
