+++ lib/main.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:gocast_mobile/model/user_model.dart';
+import 'package:gocast_mobile/model/user/user_state_model.dart';
 import 'package:gocast_mobile/viewModels/user_viewmodel.dart';
 import 'package:gocast_mobile/views/home_view.dart';
 import 'package:gocast_mobile/views/login_view.dart';
@@ -20,6 +20,8 @@ void main() {
   );
 }
 
+final scaffoldMessengerKey = GlobalKey<ScaffoldMessengerState>();
+
 class App extends ConsumerWidget {
   const App({super.key});
 
@@ -27,7 +29,16 @@ class App extends ConsumerWidget {
   Widget build(BuildContext context, WidgetRef ref) {
     final userState = ref.watch(userStateProvider);
 
+    if (userState.error != null) {
+      WidgetsBinding.instance.addPostFrameCallback((_) {
+        scaffoldMessengerKey.currentState!.showSnackBar(
+          SnackBar(content: Text('Error: ${userState.error}')),
+        );
+      });
+    }
+
     return MaterialApp(
+      scaffoldMessengerKey: scaffoldMessengerKey,
       title: 'gocast',
       debugShowCheckedModeBanner: false,
       theme: ThemeData(
