+++ lib/model/course/bookmark_model.dart
@@ -1,0 +1,13 @@
+class Bookmark {
+  Bookmark({
+    required this.id,
+    required this.userId,
+    required this.title,
+    required this.url,
+  });
+
+  int id;
+  int userId;
+  String title;
+  String url;
+}
