+++ lib/model/course/course_model.dart
@@ -1,0 +1,9 @@
+class Course {
+  Course({
+    required this.id,
+    required this.name,
+  });
+
+  int id;
+  String name;
+}
