+++ lib/model/user/user_settings_model.dart
@@ -1,0 +1,19 @@
+class UserSetting {
+  UserSetting({
+    required this.id,
+    required this.userId,
+    required this.type,
+    required this.value,
+  });
+
+  int id;
+  int userId;
+  UserSettingType type;
+  String value;
+}
+
+enum UserSettingType {
+  preferredName,
+  greeting,
+  customPlaybackSpeeds,
+}
